requirements.txt
omegaconf==2.3.0
Pillow==10.1.0
einops==0.7.0
git+https://github.com/tatsy/torchmcubes.git
trimesh==4.0.5
rembg
huggingface-hub>=0.23.2
transformers
imageio[ffmpeg]
gradio
pymeshlab
xatlas==0.0.9
moderngl==5.10.0
flask
flask-cors
werkzeug
pillow
torch
rembg
trimesh